adapted Dijstra and A* to new Graph structure

+++ src/application/graphNavigation/Navigator.java
@@ -4,5 +4,5 @@ interface Navigator {
 
     int INFINITE = Integer.MAX_VALUE;
 
-    void calculateShortestWay(Graph g, long startNode, long targetNode);
+    void calculateShortestWay(Graph g, int startNode, int targetNode);
 }
